refactor: with hook
