+++ src/components/atom_word_card_add_tag_button/index.tsx
@@ -3,22 +3,17 @@ import StyledTagButtonAtom from '@/atoms/StyledTagButton'
 import StyledTextField from '@/atoms/StyledTextField'
 import { FC, Fragment, useCallback, useState } from 'react'
 import CheckIcon from '@mui/icons-material/Check'
-import { usePutWord } from '@/hooks/words/use-put-word.hook'
-import { useRecoilCallback } from 'recoil'
-import { wordsFamily } from '@/recoil/words/words.state'
 import StyledDialog from '@/organisms/StyledDialog'
 import { DialogContent, DialogTitle } from '@mui/material'
 import { useDynamicFocus } from '@/hooks/use-dynamic-focus.hook'
-import { semesterDetailsFamily } from '@/recoil/words/semesters.state'
+import { usePutWordTagAdded } from '@/hooks/words/use-put-word-tag-added.hook'
 
 interface Props {
   wordId: string
 }
 const WordCardAddTagButton: FC<Props> = ({ wordId }) => {
   const [isAddingMode, setAddingMode] = useState(false)
-  const [loading, setLoading] = useState(false)
   const [input, setInput] = useState(``)
-  const onPutWord = usePutWord(wordId)
 
   const onClickOpen = useCallback(async () => setAddingMode(true), [])
   const [inputRef, onClickOpenWithFocus] = useDynamicFocus(onClickOpen)
@@ -26,37 +21,8 @@ const WordCardAddTagButton: FC<Props> = ({ wordId }) => {
     setInput(``)
     setAddingMode(false)
   }, [])
-
-  /** onApply depends on onResetInput after successful api call*/
-  const onApply = useRecoilCallback(
-    ({ snapshot, set }) =>
-      async () => {
-        try {
-          setLoading(true)
-
-          const wordData = await snapshot.getPromise(wordsFamily(wordId))
-          if (!wordData) throw new Error(`Word not found`)
-
-          const tagSet = new Set(wordData.tags)
-          tagSet.add(input)
-          await onPutWord({ tags: Array.from(tagSet) })
-
-          // Add the newly created tag to the semester details
-          const semesterDetails = await snapshot.getPromise(
-            semesterDetailsFamily(wordData.semester),
-          )
-          set(semesterDetailsFamily(wordData.semester), {
-            ...semesterDetails,
-            tags: Array.from(new Set([...semesterDetails.tags, input])),
-          })
-
-          onResetInput()
-        } finally {
-          setLoading(false)
-        }
-      },
-    [input, onPutWord, onResetInput],
-  )
+  const [loading, onAddTag] = usePutWordTagAdded(wordId, onResetInput)
+  const onClick = useCallback(() => onAddTag(input), [input, onAddTag])
 
   return (
     <Fragment>
@@ -81,7 +47,7 @@ const WordCardAddTagButton: FC<Props> = ({ wordId }) => {
                 right: (
                   <StyledIconButtonAtom
                     jsxElementButton={<CheckIcon fontSize="small" />}
-                    onClick={onApply}
+                    onClick={onClick}
                     isDisabled={loading || input.length === 0}
                   />
                 ),
